refactor into azureml.py
