+++ training/pyproject.toml
@@ -65,7 +65,7 @@ dependencies = [
 ]
 
 optional-dependencies.all = [  ]
-optional-dependencies.azure = [ "azure-ai-ml", "azure-identity" ]
+optional-dependencies.azure = [ "azure-ai-ml", "azure-identity", "strenum" ]
 optional-dependencies.dev = [
   "anemoi-training[docs,tests,all]",
   "pre-commit>=3.3.3",
