+++ training/src/anemoi/training/schemas/diagnostics.py
@@ -253,7 +253,6 @@ class WandbSchema(BaseModel):
 
 
 class MlflowSchema(BaseModel):
-
     enabled: bool
     "Use MLflow logger."
     offline: bool
@@ -285,12 +284,14 @@ class MlflowSchema(BaseModel):
     "Maximum number of hpParams to be logged with mlflow"
     use_azure: bool = False
     "If true, log to azure ml workspace server"
+    aml_identity: str | None = None
+    "Type of identity to use for logging in with Azure ML."
     aml_resource_group: str | None = None
     "If using AML to log with MLFlow, name of the resource group"
     aml_workspace_name: str | None = None
     "If using AML to log with MLFlow, name of the workspace"
     aml_subscription_id: str | None = None
-    "If using AML to log with MLFlow, subscription ID" 
+    "If using AML to log with MLFlow, subscription ID"
 
     @root_validator(pre=True)
     def clean_entity(cls: type["MlflowSchema"], values: dict[str, Any]) -> dict[str, Any]:  # noqa: N805
